Moving classes of "display" modules in new package (PAPI Placeholder, Motd, Signs, Scoreboard)
