+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/MissileWars.java
@@ -22,10 +22,10 @@ import co.aikar.commands.PaperCommandManager;
 import de.butzlabben.missilewars.commands.*;
 import de.butzlabben.missilewars.configuration.Config;
 import de.butzlabben.missilewars.game.GameManager;
-import de.butzlabben.missilewars.game.misc.MissileWarsPlaceholder;
+import de.butzlabben.missilewars.displays.MissileWarsPlaceholder;
 import de.butzlabben.missilewars.game.schematics.paste.FawePasteProvider;
 import de.butzlabben.missilewars.game.schematics.paste.Paster;
-import de.butzlabben.missilewars.game.signs.SignRepository;
+import de.butzlabben.missilewars.displays.signs.SignRepository;
 import de.butzlabben.missilewars.game.stats.StatsFetcher;
 import de.butzlabben.missilewars.initialization.FileManager;
 import de.butzlabben.missilewars.initialization.GamesInitialization;
