+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/displays/signs/SignRepository.java
@@ -16,7 +16,7 @@
  * along with MissileWars.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.butzlabben.missilewars.game.signs;
+package de.butzlabben.missilewars.displays.signs;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
