+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/game/Game.java
@@ -30,12 +30,12 @@ import de.butzlabben.missilewars.game.enums.GameResult;
 import de.butzlabben.missilewars.game.enums.GameState;
 import de.butzlabben.missilewars.game.enums.MapChooseProcedure;
 import de.butzlabben.missilewars.game.equipment.EquipmentManager;
-import de.butzlabben.missilewars.game.misc.MotdManager;
-import de.butzlabben.missilewars.game.misc.ScoreboardManager;
+import de.butzlabben.missilewars.displays.MotdManager;
+import de.butzlabben.missilewars.displays.ScoreboardManager;
 import de.butzlabben.missilewars.game.misc.TeamSpawnProtection;
 import de.butzlabben.missilewars.game.schematics.objects.Missile;
 import de.butzlabben.missilewars.game.schematics.objects.Shield;
-import de.butzlabben.missilewars.game.signs.MWSign;
+import de.butzlabben.missilewars.displays.signs.MWSign;
 import de.butzlabben.missilewars.game.stats.FightStats;
 import de.butzlabben.missilewars.game.timer.modules.EndTimer;
 import de.butzlabben.missilewars.game.timer.modules.GameTimer;
