+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/listener/SignListener.java
@@ -22,8 +22,8 @@ import de.butzlabben.missilewars.MissileWars;
 import de.butzlabben.missilewars.configuration.PluginMessages;
 import de.butzlabben.missilewars.game.Game;
 import de.butzlabben.missilewars.game.GameManager;
-import de.butzlabben.missilewars.game.signs.MWSign;
-import de.butzlabben.missilewars.game.signs.SignRepository;
+import de.butzlabben.missilewars.displays.signs.MWSign;
+import de.butzlabben.missilewars.displays.signs.SignRepository;
 import de.butzlabben.missilewars.util.MaterialHelper;
 import org.bukkit.ChatColor;
 import org.bukkit.block.Block;
